Move mkschedule to core/
Move the implementation of mkschedule to core/

This does not change any behaviour other than moving the algorithm to
the core/ directory and C++ lib, instead of being implemented directly
with cgo. This is a step towards making the planning stage shared among
the different shapes (slice, curtain, horizon etc) since most of the
behaviour is independent of the shape. This is determined after
implementing a prototype of the /curtain endpoint.
