+++ core/CMakeLists.txt
@@ -37,6 +37,7 @@ add_library(oneseismic
     src/base64.cpp
     src/geometry.cpp
     src/messages.cpp
+    src/plan.cpp
 )
 add_library(oneseismic::oneseismic ALIAS oneseismic)
 target_include_directories(oneseismic
